Adventures in AutoLayout
In this episode, our hero finds himself in the midst of a deep, dense
fog… the AutoLayout Swamp, to be precise. Within, he finds a sleeping
giant, the broken Module field, which he quickly slays with his mighty
Sword of StackOverflow. Next, he journeys south to an inn and creates a
manifest of his winnings: a smooth, worn leather notebook with the
title “EquipData.swift” on its cover.

Tune in next time for more exciting Adventures in AutoLayout!